cat_dog/requirements.txt
matplotlib==3.6.2
Pillow==9.3.0
Pillow==9.5.0
torch==1.12.0+cu113
torchvision==0.13.0+cu113
